LoongArch: Add CPU Autoplug support
CPU Autoplug is used for power saving. Loongson-3 supports CPUFreq, but
the effect is not very good (especially if there is only a single-thread
CPU-intensive program, the idle cores at the lowest frequency are still
consuming power because their clocks are still running). Therefore, we
introduce CPU Autoplug to adjust the number of online cores based on the
global CPU load.
Change-Id: I2dab8495b32993d02cabe555a78caf159ce5478f
Signed-off-by: Huacai Chen <chenhuacai@loongson.cn>